Restored functionality following Skype's new username validation system.
Username can no longer be sent in the URL, has to be a POST request with the username in the form data. They also moved the validator location slightly. Response codes are no longer just "406" and "200", now they're a range of 5-number responses. "20000" is "OK" and "40601" is "Registration validation errors", a.k.a. "Not available".

Thank you to @villamarinella for pointing this out in #acb1932!