Add Steam URL/ID Availability Checker
This script is slightly more complex than just checking headers for 200/404 like GitHub. Steam returns 200 regardless of whether there's a profile or not. The idea hit me to change it up a bit, and look at the length of the content returned by the server. The "not found" error is pretty short, generally around 4500 bytes, as supposed to what I've seen in the average user profile's length, usually averaging at around 7000 bytes. The only issue is private profiles or profiles not set up yet. For this event, I've added a check that reads the full content returned by the server, and specifically looks for the span opening for username formatting, "actual_persona_name". The error for a non-existent profile does not contain this, yet all existing profiles (to my knowledge) do. This is an effective, yet not foolproof, method of removing false positives. Currently, the numbers in parentheses are the content length, for debugging purposes. Private profiles with content lengths under 5000 bytes will be denoted by "(private)".