Add a decent reddit checker
This script relies on a .sh script which is also included here in the same folder; umac-reddit-verify.sh. That script is just a curl script to validate that "available" usernames are actually available. This script was rather simple to create, only problem was a 429 error from a non-browser user-agent, which I was disappointed to see Reddit using. Just in case anything else triggers it, I borrowed some headers from my browser to make it look good to the server when it's scanning. You'll notice the "(falsepos)" gets triggered a lot, and it takes a while to pass, that's simply because it's calling the "umac-reddit-verify.sh" script to verify the "available" username is actually good.